feat: add GA in layout.tsx and remove _document.tsx
